+++ core/src/main/java/com/lychcs/koikoi/scoring/ScoringEvent.java
@@ -1,9 +1,13 @@
 package com.lychcs.koikoi.scoring;
 
+/**
+ * Ein einzelner Wertungsschritt. Chips und Mult sind Dezimalwerte (double),
+ * damit Faktoren wie x1.5 nicht gerundet werden muessen.
+ */
 public record ScoringEvent(
     String sourceName,
-    long addedChips,
-    long addedMult,
+    double addedChips,
+    double addedMult,
     double xMult,
     int addedMon,
     int addedVoidDust
